restructured to fix communication issues in deployment
