--- QuestionsAndAnswers/database/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-test.js
-.env
