--- QuestionsAndAnswers/database/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "qa-dbms",
-  "version": "1.0.0",
-  "description": "Project-Catwalk Q&A Service - DBMS",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sereigh/Project-SDC.git"
-  },
-  "author": "sereigh",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/sereigh/Project-SDC/issues"
-  },
-  "homepage": "https://github.com/sereigh/Project-SDC#readme",
-  "dependencies": {
-    "@newrelic/native-metrics": "^6.0.0",
-    "mongoose": "^5.12.2",
-    "mongoose-sequence": "^5.3.1",
-    "newrelic": "^7.2.0"
-  }
-}
