+++ QuestionsAndAnswers/server/.dockerignore
@@ -3,6 +3,7 @@ Dockerfile
 .dockerignore
 docker-compose.yml
 test
+test.js
 .git
 .gitignore
 npm-debug.log
