+++ QuestionsAndAnswers/server/.gitignore
@@ -2,3 +2,4 @@ node_modules
 newrelic_agent.log
 newrelic.js
 .env
+test.js
