+++ QuestionsAndAnswers/server/database/schemas/Question.js
