+++ QuestionsAndAnswers/server/package.json
@@ -25,6 +25,8 @@
     "@newrelic/native-metrics": "^6.0.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "mongoose": "^5.12.2",
+    "mongoose-sequence": "^5.3.1",
     "newrelic": "^7.2.0"
   },
   "devDependencies": {
